# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ecom-ai API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Login",
        "operationId": "Auth_Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/PasswordLoginRequest"
                  },
                  {
                    "$ref": "#/components/schemas/CodeLoginRequest"
                  }
                ],
                "title": "Payload",
                "discriminator": {
                  "propertyName": "auth_method",
                  "mapping": {
                    "password": "#/components/schemas/PasswordLoginRequest",
                    "verification_code": "#/components/schemas/CodeLoginRequest"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_SessionBootstrap_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/token-refresh": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Refresh Token",
        "operationId": "AuthToken_Refresh",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "ecom_user_refresh",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ecom User Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_SessionBootstrap_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Logout",
        "operationId": "Auth_Logout",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/sessions": {
      "get": {
        "tags": [
          "authentication"
        ],
        "summary": "List My Sessions",
        "operationId": "AuthSession_ListMine",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_SessionSummary__"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "authentication"
        ],
        "summary": "Revoke Other Sessions",
        "operationId": "AuthSession_RevokeOthers",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/sessions/{session_id}": {
      "delete": {
        "tags": [
          "authentication"
        ],
        "summary": "Revoke Session",
        "operationId": "AuthSession_Revoke",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/auth/login": {
      "post": {
        "tags": [
          "admin-authentication"
        ],
        "summary": "Admin Login",
        "operationId": "AdminAuth_Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AdminMfaChallenge_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/mfa-verifications": {
      "post": {
        "tags": [
          "admin-authentication"
        ],
        "summary": "Verify Admin Mfa",
        "operationId": "AdminAuth_MfaVerify",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminMfaVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AdminBootstrap_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/token-refresh": {
      "post": {
        "tags": [
          "admin-authentication"
        ],
        "summary": "Refresh Admin Token",
        "operationId": "AdminAuthToken_Refresh",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "ecom_admin_refresh",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ecom Admin Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_SessionBootstrap_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/reauthentications": {
      "post": {
        "tags": [
          "admin-authentication"
        ],
        "summary": "Reauthenticate Admin",
        "operationId": "AdminAuth_Reauthenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminReauthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ReauthenticationResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/auth/logout": {
      "post": {
        "tags": [
          "admin-authentication"
        ],
        "summary": "Admin Logout",
        "operationId": "AdminAuth_Logout",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/auth/sessions": {
      "get": {
        "tags": [
          "admin-authentication"
        ],
        "summary": "List Admin Sessions",
        "operationId": "AdminAuthSession_ListMine",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_SessionSummary__"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/auth/sessions/{session_id}": {
      "delete": {
        "tags": [
          "admin-authentication"
        ],
        "summary": "Revoke Admin Session",
        "operationId": "AdminAuthSession_Revoke",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/me": {
      "get": {
        "tags": [
          "admin-authentication"
        ],
        "summary": "Get Admin Me",
        "operationId": "AdminMe_Get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AdminMe_"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/navigation": {
      "get": {
        "tags": [
          "admin-authentication"
        ],
        "summary": "Get Admin Navigation",
        "operationId": "AdminNavigation_Get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AdminNavigation_"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminLoginRequest": {
        "properties": {
          "identifier": {
            "type": "string",
            "maxLength": 254,
            "minLength": 1,
            "title": "Identifier"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          },
          "client": {
            "$ref": "#/components/schemas/ClientDescriptor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "identifier",
          "password",
          "client"
        ],
        "title": "AdminLoginRequest"
      },
...
      "AdminMfaChallenge": {
        "properties": {
          "challenge_id": {
            "type": "string",
            "title": "Challenge Id"
          },
          "allowed_methods": {
            "items": {
              "type": "string",
              "enum": [
                "totp",
                "recovery_code"
              ]
            },
            "type": "array",
            "title": "Allowed Methods"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "challenge_id",
          "allowed_methods",
          "expires_at"
        ],
        "title": "AdminMfaChallenge"
      },
      "AdminMfaVerificationRequest": {
        "properties": {
          "challenge_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 32,
            "title": "Challenge Id"
          },
          "method": {
            "type": "string",
            "enum": [
              "totp",
              "recovery_code"
            ],
            "title": "Method"
          },
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 6,
            "title": "Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "challenge_id",
          "method",
          "code"
        ],
        "title": "AdminMfaVerificationRequest"
      },
...
      "AdminReauthenticationRequest": {
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          },
          "method": {
            "type": "string",
            "enum": [
              "totp",
              "recovery_code"
            ],
            "title": "Method"
          },
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 6,
            "title": "Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "password",
          "method",
          "code"
        ],
        "title": "AdminReauthenticationRequest"
      },
...
      "ClientDescriptor": {
        "properties": {
          "client_type": {
            "type": "string",
            "const": "web",
            "title": "Client Type",
            "default": "web"
          },
          "device_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Device Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "device_name"
        ],
        "title": "ClientDescriptor"
      },
      "CodeLoginRequest": {
        "properties": {
          "auth_method": {
            "type": "string",
            "const": "verification_code",
            "title": "Auth Method"
          },
          "target_type": {
            "type": "string",
            "enum": [
              "phone",
              "email"
            ],
            "title": "Target Type"
          },
          "target": {
            "type": "string",
            "maxLength": 254,
            "minLength": 3,
            "title": "Target"
          },
          "verification_id": {
            "type": "string",
            "maxLength": 40,
            "minLength": 5,
            "title": "Verification Id"
          },
          "verification_code": {
            "type": "string",
            "pattern": "^[0-9]{6}$",
            "title": "Verification Code"
          },
          "client": {
            "$ref": "#/components/schemas/ClientDescriptor"
          },
          "challenge_token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Challenge Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "auth_method",
          "target_type",
          "target",
          "verification_id",
          "verification_code",
          "client"
        ],
        "title": "CodeLoginRequest"
      },
...
      "Envelope_AdminBootstrap_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminBootstrap"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[AdminBootstrap]"
      },
...
      "Envelope_AdminMe_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminMe"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[AdminMe]"
      },
      "Envelope_AdminMfaChallenge_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminMfaChallenge"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[AdminMfaChallenge]"
      },
      "Envelope_AdminNavigation_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminNavigation"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[AdminNavigation]"
      },
...
      "Envelope_ReauthenticationResult_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ReauthenticationResult"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[ReauthenticationResult]"
      },
...
      "Envelope_SessionBootstrap_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SessionBootstrap"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[SessionBootstrap]"
      },
...
      "Envelope_list_SessionSummary__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/SessionSummary"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[list[SessionSummary]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PasswordLoginRequest": {
        "properties": {
          "auth_method": {
            "type": "string",
            "const": "password",
            "title": "Auth Method"
          },
          "identifier": {
            "type": "string",
            "maxLength": 254,
            "minLength": 1,
            "title": "Identifier"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          },
          "client": {
            "$ref": "#/components/schemas/ClientDescriptor"
          },
          "challenge_token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Challenge Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "auth_method",
          "identifier",
          "password",
          "client"
        ],
        "title": "PasswordLoginRequest"
      },
...
      "ReauthenticationResult": {
        "properties": {
          "reauth_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Reauth Expires At"
          },
          "assurance_level": {
            "type": "string",
            "title": "Assurance Level"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reauth_expires_at",
          "assurance_level"
        ],
        "title": "ReauthenticationResult"
      },
...
      "ResponseMeta": {
        "properties": {
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "title": "ResponseMeta"
      },
...
      "SessionBootstrap": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserSummary"
          },
          "session": {
            "$ref": "#/components/schemas/SessionSummary"
          },
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "const": "Bearer",
            "title": "Token Type",
            "default": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "csrf_token": {
            "type": "string",
            "title": "Csrf Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user",
          "session",
          "access_token",
          "expires_in",
          "csrf_token"
        ],
        "title": "SessionBootstrap"
      },
...
      "SessionSummary": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "client_type": {
            "type": "string",
            "title": "Client Type"
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          },
          "audience": {
            "type": "string",
            "title": "Audience"
          },
          "authenticated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Authenticated At"
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "is_current": {
            "type": "boolean",
            "title": "Is Current",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "client_type",
          "device_name",
          "audience",
          "authenticated_at",
          "last_seen_at",
          "expires_at"
        ],
        "title": "SessionSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}